requirements.txt
# Flask and core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Session==0.5.0
Flask-Limiter==3.5.0

# Database drivers
PyMySQL==1.1.0
cryptography==41.0.4

# Authentication and security
PyJWT==2.8.0
Werkzeug==2.3.7
bcrypt==4.0.1
python-dotenv==1.0.0

# Discord integration
discord.py==2.3.2
requests==2.31.0
requests-oauthlib==1.3.1

# Image handling and cloud storage
Pillow==10.0.0
cloudinary==1.36.0

# Redis for caching and sessions
redis==5.0.0
hiredis==2.2.3

# WebSocket support
python-socketio==5.9.0
eventlet==0.33.3

# API documentation
flask-restx==1.2.0
marshmallow==3.20.1

# Development and debugging
python-decouple==3.8
gunicorn==21.2.0
supervisor==4.2.5

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Task queue (optional)
celery==5.3.2
flower==2.0.1

# Additional utilities
python-slugify==8.0.1
humanize==4.8.0
pytz==2023.3
validators==0.22.0
markupsafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.2
sqlalchemy==2.0.21
alembic==1.12.0

# Production server
uWSGI==2.0.22

# Monitoring and logging
sentry-sdk==1.32.0
structlog==23.1.0